Fix low resolution bug (#58)
WHAT: Fixes a bug which causes converted PDFs to be read and outputed
as 72 DPI images, no matter what the resolution is set to.

WHY: Resolution of the Imagick object need to be set before reading the
file.

HOW: Fixed by not passing the filename to the Imagick constructor upon
newing it up, as the file is being read later on anyways. No need to
read the file twice!